requirements.txt
asgiref==3.3.4
boto3==1.17.98
botocore==1.20.98
Django==3.2.3
django-cors-headers==3.7.0
django-dirtyfields==1.7.0
django-filter==2.4.0
django-sass-processor==1.0.1
django-storages==1.11.1
django-tables2==2.4.0
django-tables2-bootstrap4==2.0.0b5
gunicorn==20.1.0
jmespath==0.10.0
npm==0.1.1
optional-django==0.1.0
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
whitenoise==5.2.0
